requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
email-validator==2.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.27.0
PyJWT[crypto]>=2.8.0
python-dotenv==1.0.0
supabase==2.0.0
bcrypt==4.0.1